move recursive merge function to multi.py && simplify validity checks
