+++ dask/dataframe/core.py
@@ -4534,20 +4534,21 @@ class DataFrame(_Frame):
             other = other.to_frame()
 
         if not is_dataframe_like(other):
-            # TODO: make this check better
-            if isinstance(other, list) and all([is_dataframe_like(o) for o in other]):
-                return _merge_multi(
-                    self,
-                    other,
-                    on=on,
-                    how=how,
-                    lsuffix=lsuffix,
-                    rsuffix=rsuffix,
-                    npartitions=npartitions,
-                    shuffle=shuffle,
-                )
+            if not isinstance(other, list) or not all([is_dataframe_like(o) for o in other]):
+                raise ValueError("other must be DataFrame or list of DataFrames")
+            if how not in ["outer", "left"]:
+                raise ValueError("merge_multi only supports left or outer joins")
 
-            raise ValueError("other must be DataFrame or list of DataFrames")
+            from .multi import _recursive_pairwise_outer_join
+
+            other = _recursive_pairwise_outer_join(
+                other,
+                on=on,
+                lsuffix=lsuffix,
+                rsuffix=rsuffix,
+                npartitions=npartitions,
+                shuffle=shuffle,
+            )
 
         from .multi import merge
 
@@ -5064,97 +5065,6 @@ class DataFrame(_Frame):
             and (np.isscalar(key) or isinstance(key, tuple))
             and key in self.columns
         )
-
-
-def _merge_multi(
-    left,
-    others,
-    on=None,
-    how="left",
-    lsuffix="",
-    rsuffix="",
-    npartitions=None,
-    shuffle=None,
-):
-    # TODO: make these checks better
-    if not isinstance(others, list):
-        raise ValueError("other must be DataFrame or list of DataFrames")
-    if not all([is_dataframe_like(other) for other in others]):
-        raise ValueError("other must be DataFrame or list of DataFrames")
-    if how not in ["outer", "left"]:
-        raise ValueError("merge_multi only supports left or outer joins")
-
-    joined_others_df = _recursive_pairwise_outer_join(
-        others,
-        on=on,
-        lsuffix=lsuffix,
-        rsuffix=rsuffix,
-        npartitions=npartitions,
-        shuffle=shuffle,
-    )
-
-    return left.join(
-        joined_others_df,
-        on=on,
-        how=how,
-        lsuffix=lsuffix,
-        rsuffix=rsuffix,
-        npartitions=npartitions,
-        shuffle=shuffle,
-    )
-
-
-def _recursive_pairwise_outer_join(
-    dataframes_to_merge, on, lsuffix, rsuffix, npartitions, shuffle
-):
-    """
-    Schedule the merging of a list of dataframes in a pairwise method. This is a recursive function, that results
-    in a much more efficient scheduling of merges
-    from:
-    [A] [B] [C] [D] -> [AB] [C] [D] -> [ABC] [D] -> [ABCD]
-    to:
-    [A] [B] [C] [D] -> [AB] [CD] -> [ABCD]
-    Note that either way, n-1 merges are still required, but using a pairwise reduction it can be completed in parallel.
-    :param dataframes_to_merge: A list of Dask dataframes to be merged together on their index
-    :return: A single Dask Dataframe, comprised of the pairwise-merges of all provided dataframes
-    """
-    number_of_dataframes_to_merge = len(dataframes_to_merge)
-
-    merge_options = {
-        "on": on,
-        "lsuffix": lsuffix,
-        "rsuffix": rsuffix,
-        "npartitions": npartitions,
-        "shuffle": shuffle,
-    }
-
-    # Base case 1: just return the provided dataframe and merge with `left`
-    if number_of_dataframes_to_merge == 1:
-        return dataframes_to_merge[0]
-
-    # Base case 2: merge the two provided dataframe to be merged with `left`
-    if number_of_dataframes_to_merge == 2:
-        merged_ddf = dataframes_to_merge[0].join(
-            dataframes_to_merge[1], how="outer", **merge_options
-        )
-        return merged_ddf
-
-    # Recursive case: split the list of dfs into two ~even sizes and continue down
-    else:
-        middle_index = number_of_dataframes_to_merge // 2
-        merged_ddf = _recursive_pairwise_outer_join(
-            [
-                _recursive_pairwise_outer_join(
-                    dataframes_to_merge[:middle_index], **merge_options
-                ),
-                _recursive_pairwise_outer_join(
-                    dataframes_to_merge[middle_index:], **merge_options
-                ),
-            ],
-            **merge_options,
-        )
-        return merged_ddf
-
 
 # bind operators
 for op in [
